+++ NEWS.md
@@ -4,8 +4,11 @@
 * Upgraded to RSpec 3.3 [upstream-2]
 * Replaced TravisCI with CircleCI [upstream-3]
 * Rename Dev Seeds [upstream-4]
-* Remove Unicorn in favor of Puma, as [recommended by Heroku]. [upstream-5]
+* Remove Unicorn in favor of Puma, as [recommended by Heroku] [upstream-5]
 * Remove Gemfile.lock [upstream-6]
+* Add ctags configuration dotfile per project basis [upstream-7]
+* Rename HOST to APPLICATION_HOST [upstream-8],[upstream-9]
+* Remove comments and newlines in config files [upstream-10]
 
 [recommended by Heroku]: https://devcenter.heroku.com/changelog-items/594
 [upstream-1]: https://github.com/thoughtbot/suspenders/commit/f34ab5189300a57e14c28aaeca17bd5573080f1d
@@ -14,6 +17,10 @@
 [upstream-4]: https://github.com/thoughtbot/suspenders/commit/2cd79f5f6b0919340fe0ed4c69d83ff4f449dadb
 [upstream-5]: https://github.com/thoughtbot/suspenders/commit/c681ed379f4117dd62e64ff4e910a38e1a4cbf47
 [upstream-6]: https://github.com/thoughtbot/suspenders/commit/539b5c6e7fe68ee6ccad168ea6e267c7f12ae9bd
+[upstream-7]: https://github.com/thoughtbot/suspenders/commit/390c0c4f8164729927890a66d5ad62b26b10f9f4
+[upstream-8]: https://github.com/thoughtbot/suspenders/commit/9d27905884540e44109cc66630062101307f150a
+[upstream-9]: https://github.com/thoughtbot/suspenders/commit/f6f4869f9dbad9104597bc4380330e8311f75d08
+[upstream-10]: https://github.com/thoughtbot/suspenders/commit/f65e94d919796c353c399e7d1a814cd5bbf6c50d
 
 1.1.0 (June 23, 2015)
 * Added HAML and HAML Rails